# models/auth.go
package models

type LoginModel struct {
	Username string `json:"username" binding:"required"`
	Password string `json:"password" binding:"required"`
}


type TokenResponse struct {
	Token string `json:"token"`
}